[REF] resources: split files
