+++ resource_activity/__openerp__.py
@@ -43,19 +43,28 @@
         "data/resource_activity_data.xml",
         "data/cron.xml",
         "data/opening_hours.xml",
-        "views/resource_activity_views.xml",
         "views/opening_hours.xml",
         "views/partner_views.xml",
-        "views/sale_views.xml",
-        "views/resource_views.xml",
+        "views/product_views.xml",
+        "views/res_company_views.xml",
         "views/resource_activity_delivery_views.xml",
-        "views/terms_conditions_views.xml",
+        "views/resource_activity_lang_views.xml",
+        "views/resource_activity_registration_views.xml",
+        "views/resource_activity_theme_views.xml",
+        "views/resource_activity_type_views.xml",
+        "views/resource_activity_views.xml",
+        "views/resource_allocation_views.xml",
+        "views/resource_category_views.xml",
+        "views/resource_location_views.xml",
+        "views/resource_resource_views.xml",
         "views/sale_note_views.xml",
+        "views/sale_order_views.xml",
+        "views/terms_conditions_views.xml",
         "wizard/activity_draft_to_done.xml",
         "wizard/cancel_sale_order_wizard.xml",
         "reports/resource_activity_reports.xml",
-        "reports/activity_report.xml",
         "reports/activity_registration_report.xml",
+        "reports/activity_report.xml",
         "reports/sale_order_report.xml",
         "reports/layouts.xml",
         "views/menus.xml",
