+++ resource_activity/models/__init__.py
@@ -1,10 +1,19 @@
-from . import partner
-from . import resource_activity
-from . import resource_activity_registration
-from . import resource_activity_delivery
-from . import sale
-from . import resource
 from . import opening_hours
 from . import opening_hours_day
-from . import terms_conditions
+from . import partner
+from . import product
+from . import res_company
+from . import resource_activity
+from . import resource_activity_delivery
+from . import resource_activity_lang
+from . import resource_activity_registration
+from . import resource_activity_theme
+from . import resource_activity_type
+from . import resource_allocation
+from . import resource_available
+from . import resource_category
+from . import resource_location
+from . import resource_location_terms
+from . import sale_order
 from . import sale_note
+from . import terms_conditions
