+++ resource_activity/models/partner.py
@@ -2,7 +2,7 @@
 # Copyright 2018 Coop IT Easy SCRLfs.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import _, api, fields, models
+from openerp import api, fields, models
 
 
 class ResPartner(models.Model):
