+++ resource_activity/models/terms_conditions.py
@@ -2,7 +2,7 @@
 # Copyright 2020 Coop IT Easy SCRLfs
 #   - Vincent Van Rossem <vincent@coopiteasy.be>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-from openerp import api, fields, models
+from openerp import fields, models
 
 
 class TermsConditions(models.Model):
