+++ resource_activity/views/res_company_views.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="view_company_inherit_form2" model="ir.ui.view">
+        <field name="name">res.company.form.inherit</field>
+        <field name="inherit_id" ref="sale.view_company_inherit_form2"/>
+        <field name="model">res.company</field>
+        <field name="arch" type="xml">
+            <field name="sale_note" position="attributes">
+                <attribute name="invisible">True</attribute>
+            </field>
+            <field name="sale_note" position="after">
+                <field name="sale_note_html_id" nolabel="1" placeholder="Default terms &amp; conditions..."/>
+            </field>
+        </field>
+    </record>
+</odoo>
