+++ resource_activity/views/resource_activity_type_views.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+
+    <record id="resource_activity_type_tree" model="ir.ui.view">
+        <field name="name">resource.activity.type.tree</field>
+        <field name="model">resource.activity.type</field>
+        <field name="arch" type="xml">
+            <tree string="Resource activity type" delete="false">
+                <field name="name"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="view_resource_activity_type_form" model="ir.ui.view">
+        <field name="name">resource.activity.type.form</field>
+        <field name="model">resource.activity.type</field>
+        <field name="arch" type="xml">
+            <form string="Resource activity type">
+                <sheet>
+                    <div class="oe_button_box" name="button_box">
+                        <button class="oe_stat_button" type="object" name="toggle_active" icon="fa-archive">
+                            <field name="active" widget="boolean_button" options='{"terminology": "archive"}'/>
+                        </button>
+                    </div>
+                    <group>
+                        <group>
+                            <field name="name"/>
+                            <field name="location_ids" widget="many2many_tags"/>
+                        </group>
+                        <group>
+                            <!-- not used -->
+                            <field name="code" invisible="1"/>
+                            <field name="analytic_account" invisible="1"/>
+                        </group>
+                    </group>
+                    <notebook>
+                        <page name="products" string="Product">
+                            <field name="product_ids">
+                                <tree>
+                                    <field name="name"/>
+                                    <field name="default_code"/>
+                                    <field name="list_price"/>
+                                    <field name="uom_id"/>
+                                </tree>
+                            </field>
+                        </page>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="action_resource_activity_type" model="ir.actions.act_window">
+        <field name="name">Resource activity type</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">resource.activity.type</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+    </record>
+
+</odoo>
