+++ resource_activity/views/resource_allocation_views.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="view_resource_allocation_activity_form" model="ir.ui.view">
+        <field name="name">resource.allocation.form</field>
+        <field name="model">resource.allocation</field>
+        <field name="inherit_id" ref="resource_planning.view_resource_allocation_form"/>
+        <field name="arch" type="xml">
+            <field name="location" position="after">
+                <field name="activity_id"/>
+                <field name="activity_registration_id"/>
+            </field>
+        </field>
+    </record>
+</odoo>
