+++ resource_activity/views/resource_category_views.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2021 Coop IT Easy
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="view_resource_category_sale_form" model="ir.ui.view">
+        <field name="name">resource.category.form</field>
+        <field name="model">resource.category</field>
+        <field name="inherit_id" ref="resource_planning.view_resource_category_form"/>
+        <field name="arch" type="xml">
+            <page name="resources" position="after">
+                <page name="products" string="Product">
+                    <field name="product_ids">
+                        <tree>
+                            <field name="name"/>
+                            <field name="default_code"/>
+                            <field name="list_price"/>
+                            <field name="uom_id"/>
+                        </tree>
+                    </field>
+                </page>
+            </page>
+        </field>
+    </record>
+</odoo>
