+++ resource_activity/views/resource_location_views.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="resource_location_form" model="ir.ui.view">
+            <field name="name">resource.location.form</field>
+            <field name="model">resource.location</field>
+            <field name="inherit_id" ref="resource_planning.resource_location_form"/>
+            <field name="arch" type="xml">
+                <page name="users" position="after">
+                    <page name="guides" string="Guides">
+                        <field name="guides">
+                            <tree delete="false" create="false">
+                                <field name="name"/>
+                            </tree>
+                        </field>
+                    </page>
+                    <page name="trainers" string="Trainers">
+                        <field name="trainers">
+                            <tree delete="false" create="false">
+                                <field name="name"/>
+                            </tree>
+                        </field>
+                    </page>
+                    <page name="opening_hours" string="Opening Hours">
+                        <field name="opening_hours_ids">
+                            <tree delete="false" create="false">
+                                <field name="name"/>
+                                <field name="start"/>
+                                <field name="end"/>
+                                <field name="is_holiday"/>
+                            </tree>
+                        </field>
+                    </page>
+                    <page name="terms_conditions" string="Terms and Conditions">
+                        <field name="terms_ids">
+                            <tree editable="bottom">
+                                <field name="activity_type_id"/>
+                                <field name="terms_id"/>
+                                <field name="note_id"/>
+                                <field name="bike_number_display"/>
+                            </tree>
+                        </field>
+                    </page>
+                </page>
+            </field>
+        </record>
+    </data>
+</odoo>
