+++ resource_activity/views/resource_resource_views.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="resource_planning_resource_form_inherit" model="ir.ui.view">
+        <field name="name">resource.resource.form</field>
+        <field name="model">resource.resource</field>
+        <field name="inherit_id" ref="resource_planning.resource_planning_resource_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='allocations']" position="replace">
+                <field name="allocations">
+                    <tree delete="false" create="false">
+                        <field name="partner_id"/>
+                        <field name="date_start"/>
+                        <field name="date_end"/>
+                        <field name="date_lock"/>
+                        <field name="state"/>
+                        <field name="activity_id"/>
+                    </tree>
+                </field>
+            </xpath>
+        </field>
+    </record>
+</odoo>
