--- resource_activity/views/resource_views.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="view_resource_category_sale_form" model="ir.ui.view">
-            <field name="name">resource.category.form</field>
-            <field name="model">resource.category</field>
-            <field name="inherit_id" ref="resource_planning.view_resource_category_form"/>
-            <field name="arch" type="xml">
-                <page name="resources" position="after">
-                    <page name="products" string="Product">
-                        <field name="product_ids">
-                            <tree>
-                                <field name="name"/>
-                                <field name="default_code"/>
-                                <field name="list_price"/>
-                                <field name="uom_id"/>
-                            </tree>
-                        </field>
-                    </page>
-                </page>
-            </field>
-        </record>
-
-        <record id="view_resource_allocation_activity_form" model="ir.ui.view">
-            <field name="name">resource.allocation.form</field>
-            <field name="model">resource.allocation</field>
-            <field name="inherit_id" ref="resource_planning.view_resource_allocation_form"/>
-            <field name="arch" type="xml">
-                <field name="location" position="after">
-                    <field name="activity_id"/>
-                    <field name="activity_registration_id"/>
-                </field>
-            </field>
-        </record>
-
-        <record id="resource_planning_resource_form_inherit" model="ir.ui.view">
-            <field name="name">resource.resource.form</field>
-            <field name="model">resource.resource</field>
-            <field name="inherit_id" ref="resource_planning.resource_planning_resource_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='allocations']" position="replace">
-                    <field name="allocations">
-                        <tree delete="false" create="false">
-                            <field name="partner_id"/>
-                            <field name="date_start"/>
-                            <field name="date_end"/>
-                            <field name="date_lock"/>
-                            <field name="state"/>
-                            <field name="activity_id"/>
-                        </tree>
-                    </field>
-                </xpath>
-            </field>
-        </record>
-
-    </data>
-</odoo>
