+++ resource_activity/views/sale_order_views.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+
+        <record id="view_sale_order_activity_form" model="ir.ui.view">
+            <field name="name">sale.order.form</field>
+            <field name="model">sale.order</field>
+            <field name="inherit_id" ref="sale.view_order_form"/>
+            <field name="arch" type="xml">
+                <field name="project_id" position="after">
+                    <field name="activity_sale"/>
+                    <field name="activity_id" attrs="{'invisible':[('activity_sale','=',False)]}"/>
+                    <field name="date_start" attrs="{'invisible':[('activity_sale','=',False)]}"/>
+                    <field name="date_end" attrs="{'invisible':[('activity_sale','=',False)]}"/>
+                    <field name="location_id" attrs="{'invisible':[('activity_sale','=',False)]}"/>
+                </field>
+                <field name="note" position="attributes">
+                    <attribute name="invisible">True</attribute>
+                </field>
+                <field name="note" position="after">
+                    <field name="note_html_id" class="oe_inline"
+                           placeholder="Setup default terms and conditions in your company settings."/>
+                </field>
+                <group name="technical" position="after">
+                    <group name="activity" string="Activity" attrs="{'invisible':[('activity_sale','=',False)]}">
+                        <field name="description"/>
+                        <field name="departure"/>
+                        <field name="arrival"/>
+                        <field name="duration"/>
+                        <field name="langs" widget='many2many_tags'/>
+                        <field name="registrations_expected"/>
+                        <field name="activity_type"/>
+                        <field name="activity_theme"/>
+                        <field name="need_delivery"/>
+                        <field name="delivery_place" attrs="{'invisible':[('need_delivery','=',False)]}"/>
+                        <field name="delivery_time" attrs="{'invisible':[('need_delivery','=',False)]}"/>
+                        <field name="pickup_place" attrs="{'invisible':[('need_delivery','=',False)]}"/>
+                        <field name="pickup_time" attrs="{'invisible':[('need_delivery','=',False)]}"/>
+                    </group>
+                    <group colspan="4">
+                        <field name="booked_resources"/>
+                    </group>
+                </group>
+            </field>
+        </record>
+
+    </data>
+</odoo>
