+++ resource_planning/__openerp__.py
@@ -39,12 +39,14 @@
         "data/resource_planning_data.xml",
         "security/resource_planning_security.xml",
         "security/ir.model.access.csv",
-        "wizard/check_resource_availabilities_wizard.xml",
         "wizard/allocate_resource_wizard.xml",
-        "views/resource_planning_views.xml",
-        "views/resource_location_views.xml",
+        "wizard/check_resource_availabilities_wizard.xml",
         "views/partner_views.xml",
-        "views/res_config_view.xml",
+        "views/res_config_views.xml",
+        "views/resource_allocation_views.xml",
+        "views/resource_category_views.xml",
+        "views/resource_location_views.xml",
+        "views/resource_resource_views.xml",
         "views/actions.xml",
         "views/menus.xml",
     ],
