+++ resource_planning/views/res_config_views.xml
